Extract site settings localized fields
