+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.72",
-  "tag": "v3.4.72",
+  "version": "3.4.73",
+  "tag": "v3.4.73",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.71 <3.4.72"
+      ">=3.4.72 <3.4.73"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.71 first."
+    "message": "Update to v3.4.72 first."
   }
 }
